add unit test of AgencyServiceImpl
This is my first time doing a unit test of the service layer.  I tried to use Mockito but ran into trouble where I got null instead of a mock.  From StackOverflow I gathered that there are known problems with MockintoJUnitRunner and @InjectMocks, and it was better to simply roll my own.

The solution was to refactor AgencyServiceImpl to take an AgencyRepository injected via its constructor, then inject my own mocks right in the test class using the constructor.  Assuming I haven't broken anything downstream (i.e., Spring's ability to autowire an AgencyService) this should be the pattern to follow for testing the rest of the service tier.